API/requirements.txt
fastapi
uvicorn
psycopg[binary]
requests
PyYAML
